Verify gmail account with system account.
Now it's basically working.

Potential issue: when got unread gmail, after user read, it takes some time to sync with Google server, in this time, the reminder will fire even after user read the mail. Can this be avoided? Not sure... Another case is, if the sync is switched off, gmail server will always keep the unread mails, this app should not check when sync is OFF, is it correct behavior?